- Integer.valueOf with specification of radix of 16, rather than Integer.decode - Cast int to byte in assertEquals

+++ driver/src/test/unit/com/mongodb/util/JSONCallbackTest.java
@@ -96,7 +96,7 @@ public class JSONCallbackTest {
         assertArrayEquals(new byte[]{97, 98, 99, 100}, parsedBinary.getData());
 
         Binary parsedBinaryWithHexType = (Binary) JSON.parse(("{ \"$binary\" : \"YWJjZA==\", \"$type\" : \"80\" }"));
-        assertEquals(128, parsedBinaryWithHexType.getType());
+        assertEquals((byte) 128, parsedBinaryWithHexType.getType());
         assertArrayEquals(new byte[]{97, 98, 99, 100}, parsedBinaryWithHexType.getData());
     }
